backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
groq==0.4.1
chromadb==0.4.22
sentence-transformers==2.3.1
beautifulsoup4==4.12.3
requests==2.31.0
python-multipart==0.0.6
sqlalchemy==2.0.25
aiosqlite==0.19.0
